For full changelog read all lines
Add 3 new Model classes, Course, Degree and Project

Add Api on Cources, Degrees and Projects

Add query string search to the projects api similar to the jobs api, only difference is that search on Locations is not permitted

Add code that inserts mock data to the tables Cource, Degree and Project. This data aims to be as realistic as possible,  in one of the projects the field description also contains html elements which seems to be allowed on the web page.